+++ VFS/source/VFSGenericMemReadOnlyFile.cpp
@@ -23,13 +23,13 @@ ssize_t VFSGenericMemReadOnlyFile::Read(void *_buf, size_t _size)
         return 0;
     
     // we can only deal with cache buffer now, need another branch later
-    if(m_Pos == m_Size)
+    if(m_Pos == (long)m_Size)
         return 0;
     
     size_t to_read = MIN(m_Size - m_Pos, _size);
     memcpy(_buf, (char*)m_Mem + m_Pos, to_read);
     m_Pos += to_read;
-    assert(m_Pos <= m_Size); // just a sanity check
+    assert(m_Pos <= (long)m_Size); // just a sanity check
         
     return to_read;
 }
@@ -40,7 +40,7 @@ ssize_t VFSGenericMemReadOnlyFile::ReadAt(off_t _pos, void *_buf, size_t _size)
         return VFSError::InvalidCall;
     
     // we can only deal with cache buffer now, need another branch later
-    if(_pos < 0 || _pos > m_Size)
+    if(_pos < 0 || _pos > (long)m_Size)
         return VFSError::InvalidCall;
     
     ssize_t toread = MIN(m_Size - _pos, _size);
@@ -65,7 +65,7 @@ off_t VFSGenericMemReadOnlyFile::Seek(off_t _off, int _basis)
     
     if(req_pos < 0)
         return VFSError::InvalidCall;
-    if(req_pos > m_Size)
+    if(req_pos > (long)m_Size)
         req_pos = m_Size;
     m_Pos = req_pos;
     
@@ -93,7 +93,7 @@ bool VFSGenericMemReadOnlyFile::Eof() const
 {
     if(!IsOpened())
         return true;
-    return m_Pos == m_Size;
+    return m_Pos == (long)m_Size;
 }
 
 int VFSGenericMemReadOnlyFile::Open(int _open_flags, const VFSCancelChecker &_cancel_checker)
